+++ src/handle_int.c
@@ -6,7 +6,7 @@
 /*   By: erlazo <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/02/18 19:56:08 by erlazo            #+#    #+#             */
-/*   Updated: 2020/02/26 16:57:24 by erlazo           ###   ########.fr       */
+/*   Updated: 2020/02/27 18:06:26 by erlazo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,7 +35,7 @@ int				ft_handle_int(va_list ap, char **str, t_param *p)
 	char		*pre;
 	char		*post;
 	char		c;
-	long		num;		// just make it a long long ???? size_t ????
+	long long	num;		// just make it a long long ???? size_t ????
 	size_t		len;
 	size_t		wlen;
 	size_t		plen;
@@ -52,7 +52,11 @@ int				ft_handle_int(va_list ap, char **str, t_param *p)
 
 	printf("handle int test 1\n");
 
-		// diuUxX		// possibly also oO other things too ????	better way of doing this ????
+
+	
+
+	
+
 
 	if (p->spec == 'd' || p->spec == 'i')	// regular ints i think
 	{
@@ -81,7 +85,7 @@ int				ft_handle_int(va_list ap, char **str, t_param *p)
 		len = ft_fstrlen(tmp);
 //		printf("its a hex, tmp: |%s|\n", tmp);
 	}
-	else if (p->spec == 'b' || p->spec == 'B')		// add binary here ???
+	else if (p->spec == 'b' || p->spec == 'B')		// add binary here ???		// could put in seperate file ???
 	{
 		if (p->spec == 'b')
 			post = va_arg(ap, char*);	// can not free post, not sure why, ok so can't free because its a arg passed to func, so don't worry, just set to NULL...
